bot/requirements.txt
Flask==2.2.5
requests
beautifulsoup4
python-telegram-bot
gunicorn
